+++ MatrixPilot/flightplan-waypoints.c
@@ -24,21 +24,21 @@
 #if (FLIGHT_PLAN_TYPE == FP_WAYPOINTS)
 
 
-struct relWaypointDef { struct relative3D loc ; int flags ; struct relative3D viewpoint ; } ;
-struct waypointDef { struct waypoint3D loc ; int flags ; struct waypoint3D viewpoint ; } ;
+struct relWaypointDef { struct relative3D loc ; int16_t flags ; struct relative3D viewpoint ; } ;
+struct waypointDef { struct waypoint3D loc ; int16_t flags ; struct waypoint3D viewpoint ; } ;
 
 #include "waypoints.h"
 
 #if (SERIAL_OUTPUT_FORMAT == SERIAL_MAVLINK)
-unsigned int  number_of_waypoints = (( sizeof waypoints ) / sizeof ( struct waypointDef ));
+uint16_t  number_of_waypoints = (( sizeof waypoints ) / sizeof ( struct waypointDef ));
 #endif
 #define NUMBER_POINTS (( sizeof waypoints ) / sizeof ( struct waypointDef ))
 #define NUMBER_RTL_POINTS (( sizeof rtlWaypoints ) / sizeof ( struct waypointDef ))
 
-int waypointIndex = 0 ;
+int16_t waypointIndex = 0 ;
 
 struct waypointDef *currentWaypointSet = (struct waypointDef*)waypoints ;
-int numPointsInCurrentSet = NUMBER_POINTS ;
+int16_t numPointsInCurrentSet = NUMBER_POINTS ;
 
 struct waypointDef wp_inject ;
 unsigned char wp_inject_pos = 0 ;
@@ -79,7 +79,7 @@ struct relWaypointDef wp_to_relative(struct waypointDef wp)
 
 // In the future, we could include more than 2 waypoint sets...
 // flightplanNum is 0 for main waypoints, and 1 for RTL waypoints
-void init_flightplan ( int flightplanNum )
+void init_flightplan ( int16_t flightplanNum )
 {
 	if ( flightplanNum == 1 ) // RTL waypoint set
 	{
@@ -121,7 +121,7 @@ struct absolute3D get_fixed_origin( void )
 	struct absolute3D standardizedOrigin ;
 	standardizedOrigin.x = origin.x ;
 	standardizedOrigin.y = origin.y ;
-	standardizedOrigin.z = (long)(origin.z * 100) ;
+	standardizedOrigin.z = (int32_t)(origin.z * 100) ;
 	
 	return standardizedOrigin ;
 }
@@ -192,7 +192,7 @@ void run_flightplan( void )
 	
 	if ( desired_behavior._.altitude )
 	{
-		if ( abs(IMUheight - goal.height) < ((int) HEIGHT_MARGIN ))
+		if ( abs(IMUheight - goal.height) < ((int16_t) HEIGHT_MARGIN ))
 			next_waypoint() ;
 	}
 	else
